requirements.txt
aisuite[all]
gradio
python-dotenv